Readded blake2 precompile.
squash
