+++ blake2/Cargo.toml
@@ -1,20 +1,16 @@
 [package]
 name = "ewasm-precompile-blake2"
 version = "0.2.0"
-authors = ["Grant Wuerker <gwuerker@gmail.com>"]
+authors = ["Alex Beregszaszi <alex@rtfs.hu>"]
 license = "Apache-2.0"
 repository = "https://github.com/ewasm/ewasm-precompiles"
-description = "Ethereum Blake2 compression function precompile in Rust"
+description = "Ethereum Blake2 precompile in Rust"
 publish = false
 edition = "2018"
 
 [dependencies]
-byteorder = "1.3.4"
+blake2 = "0.7"
 ewasm_api = "0.9"
-arr_macro = "0.1.3"
-
-[dev-dependencies]
-hex = "0.3.1"
 
 [lib]
 crate-type = ["cdylib"]
