+++ blake2f/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "ewasm-precompile-blake2f"
+version = "0.2.0"
+authors = ["Grant Wuerker <gwuerker@gmail.com>"]
+license = "Apache-2.0"
+repository = "https://github.com/ewasm/ewasm-precompiles"
+description = "Ethereum Blake2 compression function precompile in Rust"
+publish = false
+edition = "2018"
+
+[dependencies]
+byteorder = "1.3.4"
+ewasm_api = "0.9"
+arr_macro = "0.1.3"
+
+[dev-dependencies]
+hex = "0.3.1"
+
+[lib]
+crate-type = ["cdylib"]
